Refactor src-tauri/src/agent/workflow into sub-modules
This commit splits the monolithic src-tauri/src/agent/workflow.rs file into
a modular directory structure within src-tauri/src/agent/workflow/.

The new modules are:
- start.rs: Workflow initialization logic
- pause_resume.rs: Workflow state transitions
- cancel.rs: Workflow termination and cancellation logic
- tool.rs: Tool result processing logic
- mod.rs: Module re-exports

No functional changes or API modifications were made.
