+++ src-tauri/src/agent/workflow/pause_resume.rs
@@ -1,0 +1,81 @@
+use crate::agent::state::AgentSession;
+use crate::mcp::MCPServiceProxyManager;
+use crate::repositories::session_repository::SessionRepository;
+use crate::repositories::SessionStatus;
+use std::collections::HashMap;
+use std::sync::atomic::Ordering;
+use std::sync::Arc;
+use tauri::AppHandle;
+use tokio::sync::RwLock;
+/// Pause a running workflow
+pub async fn pause_workflow(
+    session_repo: &Arc<dyn SessionRepository>,
+    active_sessions: &Arc<RwLock<HashMap<String, AgentSession>>>,
+    app_handle: &AppHandle,
+    session_id: String,
+) -> Result<(), String> {
+    crate::agent::lifecycle::update_session_status(
+        session_repo,
+        active_sessions,
+        app_handle,
+        &session_id,
+        SessionStatus::Paused,
+    )
+    .await?;
+
+    let mut active = active_sessions.write().await;
+    if let Some(session) = active.get_mut(&session_id) {
+        session.is_running = false;
+    }
+
+    log::info!("Paused workflow for session: {}", session_id);
+    Ok(())
+}
+
+/// Resume a paused workflow
+pub async fn resume_workflow(
+    session_repo: &Arc<dyn SessionRepository>,
+    active_sessions: &Arc<RwLock<HashMap<String, AgentSession>>>,
+    proxy_manager: &Arc<MCPServiceProxyManager>,
+    app_handle: &AppHandle,
+    session_id: String,
+) -> Result<(), String> {
+    {
+        let active = active_sessions.read().await;
+        if let Some(session) = active.get(&session_id) {
+            session.cancel_pending.store(false, Ordering::SeqCst);
+        }
+    }
+
+    // Ensure cache is initialized before resuming (lazy load if needed, preserve if exists)
+    crate::agent::lifecycle::ensure_cache_initialized(active_sessions, &session_id).await?;
+
+    crate::agent::lifecycle::update_session_status(
+        session_repo,
+        active_sessions,
+        app_handle,
+        &session_id,
+        SessionStatus::Busy,
+    )
+    .await?;
+
+    let mut active = active_sessions.write().await;
+    if let Some(session) = active.get_mut(&session_id) {
+        session.is_running = true;
+    }
+    drop(active); // Drop lock before async call
+
+    log::info!("Resumed workflow status for session: {}", session_id);
+
+    // Trigger LLM to pick up where it left off
+    crate::agent::llm::request_llm_completion(
+        session_repo,
+        active_sessions,
+        proxy_manager,
+        app_handle,
+        session_id,
+    )
+    .await?;
+
+    Ok(())
+}
